Add the TKRangeSelect class
This class changes a value within a range. Given a minimum and a
maximum, it increments the variable by one until it reaches the max,
after which, it wraps around to the minimum again. The default range is
from 0 to the number of children minus 1 to make it easy to use with the
TKSwitch class.